refactor: moved builtin out of plugin since it has nothign to do with plugin
