+++ jac/jaclang/cli/cli.py
@@ -18,10 +18,10 @@ from jaclang.compiler.constant import Constants
 from jaclang.compiler.passes.main.pyast_load_pass import PyastBuildPass
 from jaclang.compiler.passes.main.schedules import py_code_gen_build, py_code_gen_typed
 from jaclang.compiler.program import JacProgram
+from jaclang.runtimelib.builtin import dotgen
 from jaclang.runtimelib.constructs import WalkerArchitype
 from jaclang.runtimelib.context import ExecutionContext
 from jaclang.runtimelib.machine import JacMachine
-from jaclang.runtimelib.plugin.builtin import dotgen
 from jaclang.runtimelib.plugin.feature import JacCmd as Cmd
 from jaclang.runtimelib.plugin.feature import JacFeature as Jac
 from jaclang.utils.helpers import debugger as db
